Disable page for unregistering from events if the user has been accepted

+++ selvbetjening/frontend/eventportal/views.py
@@ -3,9 +3,8 @@
 from django.core.urlresolvers import reverse
 from django.contrib.auth.decorators import login_required
 from django.db.models.aggregates import Count
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, HttpResponseForbidden, Http404
 from django.shortcuts import render, get_object_or_404
-from django.template import Context, Template
 from django.utils.translation import ugettext as _
 from django.contrib import messages
 
@@ -34,13 +33,18 @@ def event_detail(request,
           event_pk,
           template_name='eventportal/event.html'):
 
-    instance = get_object_or_404(Event, pk=event_pk)
+    event = get_object_or_404(Event, pk=event_pk)
+
+    try:
+        attendee = Attend.objects.get(event=event, user=request.user)
+    except Attend.DoesNotExist:
+        attendee = None
 
     # Attendee stats
 
     attendee_stats = {}
 
-    for item in instance.attendees.values('state').annotate(count=Count('pk')):
+    for item in event.attendees.values('state').annotate(count=Count('pk')):
         attendee_stats[item['state']] = item['count']
 
     attendee_stats['accepted'] = attendee_stats.get('accepted', 0) + attendee_stats.get('attended', 0)
@@ -51,17 +55,18 @@ def event_detail(request,
 
     # Options stats
 
-    public_options = dynamic_statistics(instance)
+    public_options = dynamic_statistics(event)
 
     return render(request,
                   template_name,
                   {
-                      'event': instance,
-                      'is_attendee': instance.is_attendee(request.user),
+                      'event': event,
+                      'is_attendee': attendee is not None,
+                      'attendee': attendee,
                       'attendee_stats': attendee_stats,
                       'public_options': public_options,
-                      'waiting_attendees': instance.attendees.filter(state=AttendState.waiting).select_related('user'),
-                      'confirmed_attendees': instance.attendees.exclude(state=AttendState.waiting).select_related('user')
+                      'waiting_attendees': event.attendees.filter(state=AttendState.waiting).select_related('user'),
+                      'confirmed_attendees': event.attendees.exclude(state=AttendState.waiting).select_related('user')
                   })
 
 
@@ -135,7 +140,10 @@ def event_register(request,
 @suspend_automatic_attendee_price_updates
 def event_unregister(request, event):
 
-    attendee = Attend.objects.get(user=request.user, event=event)
+    attendee = get_object_or_404(Attend, user=request.user, event=event)
+
+    if attendee.state != AttendState.waiting:
+        return HttpResponseForbidden()
 
     if request.method == 'POST':
         attendee.delete()
